Project 1
The aim of this project is to predict the housing prices in California using a Linear Regression Analysis. This is done using the python programming language through which a machine learning model is designed and allowed to use the dataset to train itself and come out with the predicted prices which are then compared to the current prices.